+++ R/date_to_unix.R
@@ -1,0 +1,4 @@
+date_to_unix <- function(Date) {
+  posixct <- as.POSIXct(as.Date(Date, origin = "1970-01-01"))
+  trunc(as.numeric(posixct))
+}